requirements.txt

django
psycopg2
djangorestframework
flake8
sphinx
markdown
httpie